requirements.txt
flask
flask-cors
python-dotenv
supabase
bcrypt
pyjwt
qrcode
pillow
gunicorn